==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_microbiome_composition_analysis.R")
source(file = "R/05_metabolites_PCA.R")
source(file = "R/06_behaviors_vs_treatment.R")
source(file = "R/07_analysis_v.R")
source(file = "R/07_GIsymptoms_barplots.R")
source(file = "R/08_aberrant_behavior_vs_taxa.R")
source(file = "R/09_GIsymptoms_vs_taxa_boxplots.R")
source(file = "R/10_analysis_vi.R")
source(file = "R/11_metabolites_vs_bacterias.R")
rmarkdown::render(input = "doc/Group10_presentation.Rmd")
